Fix misused environment variable name
Action was attempting to find the GitHub home directory using the `GITHUB_HOME` environment variable, but the correct one is simple `HOME`. A simple change was made to the Test Run workflow in which the image is pushed on behalf of matootie, instead of a value specified in a secret.

.github/workflows/test.yml | 2 +-
dist/index.js              | 2 +-
index.js                   | 2 +-
3 files changed, 3 insertions(+), 3 deletions(-)